+++ Sources/ComposerApp/Views/EditingStage.swift
@@ -471,7 +471,9 @@ struct EditingStage: View {
   }
 
   private func commitLabel() {
-    interaction.text = labelDraft
+    if labelDraft != interaction.text {
+      board.setText(card.id, labelDraft)
+    }
     board.endEditing(card.id)
   }